Made 5 different pages, for 4 pages for features,pricing,architecture and how it works page and 1 for navigation. Did some required changes in layout page also
