+++ apps/web/app/layout.tsx
@@ -24,7 +24,10 @@ export default function RootLayout({ children }: { children: React.ReactNode }) 
   return (
     <html lang="en" className={`${inter.variable} ${jetbrainsMono.variable}`}>
       <body className="font-sans antialiased bg-primary text-text">
-        <Providers>{children}</Providers>
+        <Providers>
+          <Navbar /> 
+          {children}
+        </Providers>
       </body>
     </html>
   );
